deleted algae and coral subsystems
